*: Rename xapi to distsql
distsql is a more clear name.
