+++ store/tikv/mock-tikv/cop_handler.go
@@ -20,14 +20,14 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/juju/errors"
 	"github.com/pingcap/kvproto/pkg/coprocessor"
+	"github.com/pingcap/tidb/distsql"
+	"github.com/pingcap/tidb/distsql/xeval"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/mysql"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/terror"
 	"github.com/pingcap/tidb/util/codec"
 	"github.com/pingcap/tidb/util/types"
-	"github.com/pingcap/tidb/xapi"
-	"github.com/pingcap/tidb/xapi/xeval"
 	"github.com/pingcap/tipb/go-tipb"
 )
 
@@ -194,7 +194,7 @@ func (h *rpcHandler) getRowsFromSelectReq(ctx *selectContext) ([]*tipb.Row, erro
 		if col.GetPkHandle() {
 			continue
 		}
-		ctx.colTps[col.GetColumnId()] = xapi.FieldTypeFromPBColumn(col)
+		ctx.colTps[col.GetColumnId()] = distsql.FieldTypeFromPBColumn(col)
 	}
 
 	kvRanges, desc := h.extractKVRanges(ctx)
@@ -435,7 +435,7 @@ func (h *rpcHandler) setColumnValueToCtx(ctx *selectContext, handle int64, row m
 			}
 		} else {
 			data := row[colID]
-			ft := xapi.FieldTypeFromPBColumn(col)
+			ft := distsql.FieldTypeFromPBColumn(col)
 			datum, err := tablecodec.DecodeColumnValue(data, ft)
 			if err != nil {
 				return errors.Trace(err)
